// types/groupBuy.ts
export interface IGroupBuy {
  Id: string;
  Id_user: number;
  Brand: string;
  TelegramUsername: number | string;
  ContactPerson: string;
  Category: string;
  Date: string;
  Details: string;
  ExpectedDelivery: string;
  Item: string;
  Name: string;
  OrderClosing: string;
  PickupLocation: string;
}